1. version updated to 2.5 2. amount validation added on unable to self award transaction amount

+++ app/src/main/java/com/onelink/nrlp/android/utils/ValidationUtils.kt
@@ -126,7 +126,7 @@ object ValidationUtils {
     }
 
     fun isAmountValid(string: String): Boolean {
-        return string.isNotEmpty() && string.toInt() > 0
+        return string.isNotEmpty() && string.toDouble() > 0
     }
 
     fun isTransactionNoValid(string: String): Boolean {
